# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-distance3d",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — Euclidean distance between two 3D points.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. `dx = x2 - x1`, `dy = y2 - y1`, `dz = z2 - z1` via `srvcs-floatsubtract`;\n2. square each via `srvcs-floatmultiply`;\n3. `sum = (dx2 + dy2) + dz2` by chaining two `srvcs-floatadd` calls;\n4. `result = sqrt(sum)` via `srvcs-sqrt`.\n\nIf a dependency is unreachable it reports itself degraded (`503`); if a\ndependency rejects an input it forwards the `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Distance3dResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected an input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Distance3dResponse": {
        "type": "object",
        "required": [
          "x1",
          "y1",
          "z1",
          "x2",
          "y2",
          "z2",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double"
          },
          "x1": {
            "type": "object"
          },
          "x2": {
            "type": "object"
          },
          "y1": {
            "type": "object"
          },
          "y2": {
            "type": "object"
          },
          "z1": {
            "type": "object"
          },
          "z2": {
            "type": "object"
          }
        }
      },
      "EvalRequest": {
        "type": "object",
        "required": [
          "x1",
          "y1",
          "z1",
          "x2",
          "y2",
          "z2"
        ],
        "properties": {
          "x1": {
            "type": "object",
            "description": "X coordinate of the first point."
          },
          "x2": {
            "type": "object",
            "description": "X coordinate of the second point."
          },
          "y1": {
            "type": "object",
            "description": "Y coordinate of the first point."
          },
          "y2": {
            "type": "object",
            "description": "Y coordinate of the second point."
          },
          "z1": {
            "type": "object",
            "description": "Z coordinate of the first point."
          },
          "z2": {
            "type": "object",
            "description": "Z coordinate of the second point."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}